Move multiple tables test to CompactionJobCompletionTest
